[cores] all: Generate OKAY response when accessing an invalid address
Responding with SLVERR causes trouble if there's an AXI downsizer in the path.
Transactions are split into multiple smaller ones, but if one of them generates
an error, it will propagate to the original transaction. This can happen, for example,
when reading 16 bytes from the end of a region with a size that is not a multiple of
16 bytes.